xtask: refactor
