+++ xtask/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 anyhow.workspace = true
 culpa.workspace = true
 derive-where.workspace = true
+pathdiff = "0.2.3"
 strum = { version = "0.28.0", features = ["derive"] }
 
 [dependencies.clap]
